+++ components/HeaderNav/HeaderNav.tsx
@@ -1,10 +1,7 @@
 import React, { useState } from "react";
 import { Button, CircularProgress, Flex, HStack, Text, Img } from "@chakra-ui/react";
-import Web3 from "web3";
-import Web3Modal from "web3modal";
-import WalletConnectProvider from "@walletconnect/web3-provider";
 import { useRouter } from "next/router";
-import { UserData } from "../../utils/interfaces";
+import { handleWalletConnect } from "../../utils/walletUtils";
 
 const HeaderNav = (props) => {
   const whichPage = props.whichPage;
@@ -14,64 +11,6 @@ const HeaderNav = (props) => {
   const [isSubmitted, setIsSubmitted] = useState(false);
   const [loaded, setLoaded] = useState(false);
   const router = useRouter();
-
-  const getUserByWallet = async (userWallet) => {
-    let lowerCaseWallet = userWallet.toLowerCase();
-    const res = await fetch(`/api/users/wallet/${lowerCaseWallet}`);
-
-    const data: any = await res.json();
-
-    console.log("RETRIEVE USER BY WALLET YO");
-    return data;
-  };
-  const handleWalletConnect = async () => {
-    try {
-    const web3Modal = new Web3Modal({
-      disableInjectedProvider: false,
-      network: "rinkeby",
-      cacheProvider: false,
-      providerOptions: {
-        walletconnect: {
-          package: WalletConnectProvider,
-          options: {
-            rpc: {
-              1: "https://eth-rinkeby.alchemyapi.io/v2/QtLM8rW9nB6DobDu8KQx-7fYMS2rBlky",
-            },
-          },
-        },
-      },
-    });
-    web3Modal.clearCachedProvider();
-    const provider = await web3Modal.connect();
-    const web3 = new Web3(provider);
-    const accounts = await web3.eth.getAccounts();
-    const currAccount = accounts[0];
-
-    setIsSubmitted(true);
-      let userData: UserData = await getUserByWallet(currAccount);
-      console.log(userData);
-      if (userData && userData.userName && userData.userWallet) {
-        if (router.query?.view == 'public' && userPage && userPage.userName) {
-          await router.push(`/${userPage.userName}`);
-          router.reload(); // reloads the profile page after changing the shallow route
-        } else if (router.query?.view == 'public' && userData && userData.userName) {
-          await router.push(`/${userData.userName}`);
-          router.reload(); // reloads the profile page after changing the shallow route
-        } else {
-          router.reload(); // reloads the profile page upon login
-        }
-        
-        setIsSubmitted(false);
-      } else {
-        console.log("No profile, pls create one...");
-        router.push("/steps");
-      }
-    } catch (err) {
-      console.log("error: ", err);
-      router.push("/steps");
-      setLoaded(true);
-    }
-  };
 
   return (
     <Flex
@@ -100,7 +39,7 @@ const HeaderNav = (props) => {
              height={"52px"}
              color={"#062B2A"}
              backgroundColor={"#C7F83C"}
-             onClick={handleWalletConnect}
+             onClick={() => handleWalletConnect(userPage, setIsSubmitted, setLoaded, router)}
            >
              Connect Wallet{" "}
              {isSubmitted ? (
